long changed to int

+++ src/buildHelper.h
@@ -1,12 +1,11 @@
 #pragma once
-#include "node.h"
 #include "edge.h"
+#include "node.h"
 
 using namespace std;
 
-class buildHelper{
-    public:
-        void insertNode(vector<string> line);
-        void insertEdge(vector<string> otherInfo, vector<long> authors);
-
+class buildHelper {
+public:
+  void insertNode(vector<string> line);
+  void insertEdge(vector<string> otherInfo, vector<int> authors);
 };
